Pull from master

+++ simutrans/ai/sqai_rail/industry_manager.nut
@@ -15,8 +15,12 @@ class industry_link_t
 	state = 0
 	lines = null   // array<line_x>
 
-	double_ways_count = 0 // count of double way build
-	double_ways_build = 0 // double way build: 0 = no ; 1 = yes
+	double_ways_count		= 0 // count of double way build
+	double_ways_build		= 0 // double way build: 0 = no ; 1 = yes
+	optimize_way_line		= 0 // is way line optimize: 0 = no ; 1 = yes
+	destroy_line_month	= null // test month save
+	line_way_speed			= 0 // save way speed for line
+	build_line					= 0 // create line
 
 	// next check needed if ticks > next_check
 	// state == st_missing: check availability again
@@ -194,6 +198,13 @@ class industry_manager_t extends manager_t
 	function check_link_line(link, line)
 	{
 
+		// set first run as line build time
+		if ( link.build_line == 0 ) {
+  		link.build_line = world.get_time()
+		}
+
+		//
+
 		local  print_message_box = 0
 
 		dbgprint("Check line " + line.get_name())
@@ -201,15 +212,91 @@ class industry_manager_t extends manager_t
 		// find convoy
 		local cnv = null
 		local cnv_count = 0
+		local cnv_max_speed = 0
 		{
 			local list = line.get_convoy_list()
 			cnv_count = list.get_count()
 			if (cnv_count == 0) {
 				// 0 convoy destroy line
-				if ( line.get_owner().nr == our_player.nr ) { destroy_line(line) }
+				if ( line.get_owner().nr == our_player.nr ) {
+					destroy_line(line)
+					sleep()
+				}
 				return
 			}
-			cnv = list[0]
+			for ( local i = 0; i < cnv_count; i++ ) {
+				if ( !list[i].is_withdrawn() ) {
+					cnv = list[i]
+					break
+				}
+			}
+			// check speed from convoys
+			local a = convoy_max_speed(list[0])
+			local cnv_retired = []
+
+			// check cnv is retired
+			if ( list[0].has_obsolete_vehicles() && !list[0].is_withdrawn() ) {
+				//gui.add_message_at(our_player, " retired convoy " + list[i].get_name(), world.get_time())
+				cnv_retired.append(list[0])
+			}
+
+			local speed_count = 1
+			for ( local i = 1; i < cnv_count; i++ ) {
+				local s = convoy_max_speed(list[i])
+
+				if ( a == s ) {
+					speed_count++
+				} else if ( a < s && cnv_max_speed == 0 ) {
+					cnv_max_speed = s
+				}
+				// check cnv is retired
+				if ( list[i].has_obsolete_vehicles() && !list[i].is_withdrawn() ) {
+					gui.add_message_at(our_player, " retired convoy " + list[i].get_name(), world.get_time())
+					cnv_retired.append(list[i])
+				}
+			}
+			//gui.add_message_at(our_player, " max speed " + cnv_max_speed, world.get_time())
+
+			if ( speed_count < cnv_count ) {
+				// update convoys - retire slower convoys
+				local k = 0
+				for ( local i = 0; i < cnv_count; i++ ) {
+					if ( convoy_max_speed(list[i]) < cnv_max_speed && !list[i].is_withdrawn() ) {
+						list[i].toggle_withdraw(our_player)
+						k++
+					}
+				}
+				if ( k > 0 ) {
+					local msgtext = format(translate("%d convoys of the line %s were retired"), k, line.get_name())
+					gui.add_message_at(our_player, msgtext, world.get_time())
+
+					return
+
+				}
+			}
+
+			if ( cnv_retired.len() > 0 && cnv_retired.len() < cnv_count ) {
+				// update convoys - retire retired convoys
+				local k = 0
+				for ( local i = 0; i < cnv_retired.len(); i++ ) {
+					if ( !cnv_retired[i].is_withdrawn() ) {
+						cnv_retired[i].toggle_withdraw(our_player)
+						k++
+					}
+				}
+				if ( k > 0 ) {
+					local msgtext = format(translate("%d convoys of the line %s were retired"), k, line.get_name())
+					gui.add_message_at(our_player, msgtext, world.get_time())
+
+					return
+
+				}
+			} else if ( cnv_retired.len() == 1 && cnv_retired.len() == cnv_count ) {
+				//gui.add_message_at(our_player, "*** cnv_upgrade = 1 -> line " + line.get_name(), world.get_time())
+				// create new convoy before retire retired convoy
+				upgrade_link_line(link, line)
+				return
+			}
 		}
 
 		if ( line.get_owner().nr == our_player.nr ) {
@@ -217,8 +304,11 @@ class industry_manager_t extends manager_t
 			local profit_count = line.get_profit()
 			//if ( cnv.get_distance_traveled_total() < 3 ) { return }
 			if ( (profit_count[4] < 0 || profit_count[4] == 0) && profit_count[3] == 0 && profit_count[2] == 0 && profit_count[1] == 0 && profit_count[0] == 0 ) {
-				if ( cnv.get_distance_traveled_total() > 1 && cnv.get_distance_traveled_total() < 25 && cnv.get_loading_level() == 0 ) {
-					destroy_line(line)
+				if ( cnv.get_distance_traveled_total() > 1 && cnv.get_distance_traveled_total() < 25 && cnv.get_loading_level() == 0 && link.destroy_line_month != world.get_time().month ) {
+					local erreg = destroy_line(line)
+					if ( erreg == false ) {
+						link.destroy_line_month = world.get_time().month
+					}
 				} else {
 					//gui.add_message_at(our_player, "return cnv/line new " + line.get_name(), world.get_time())
 				}
@@ -227,8 +317,18 @@ class industry_manager_t extends manager_t
 		}
 
 		sleep()
-		if (our_player.get_current_cash() > 5000000 && cnv.get_waytype() != wt_water && cnv.get_waytype() != wt_air) {
-			local nexttile = [] //[tile_x(start.x, start.y, start.z)]
+		if ( !cnv.is_valid() ) { return }
+
+		// find route
+		local nexttile = []
+		if (cnv.get_waytype() != wt_water && cnv.get_waytype() != wt_air) {
+			local entries = cnv.get_schedule().entries
+			local start = null
+			local end = null
+			if ( entries.len() >= 2 ) {
+				start = tile_x(entries[0].x, entries[0].y, entries[0].z)
+				end = tile_x(entries[entries.len()-1].x, entries[entries.len()-1].y, entries[entries.len()-1].z)
+			}
 
 			local asf = astar_route_finder(cnv.get_waytype())
 			local result = asf.search_route([start], [end])
@@ -236,7 +336,7 @@ class industry_manager_t extends manager_t
 			// route is backward from end to start
 
 			if ("err" in result) {
-				gui.add_message_at(our_player, " ### no route found: " + result.err, start)
+				//gui.add_message_at(our_player, " ### no route found: " + result.err, start)
 				return nexttile
 			}
 			else {
@@ -248,24 +348,85 @@ class industry_manager_t extends manager_t
 				}
 				sleep()
 			}
-			// optimize way line befor build double ways
-			optimize_way_line(nexttile, cnv.get_waytype())
 		}
 
-		if (cnv.is_withdrawn()) {
+
+		if (our_player.get_current_cash() > 50000 && cnv.get_waytype() != wt_water && cnv.get_waytype() != wt_air) {
+			if ( link.optimize_way_line == 0 ) {
+				// optimize way line befor build double ways
+				optimize_way_line(nexttile, cnv.get_waytype())
+				link.optimize_way_line = 1
+			} else {
+
+			}
+		}
+
+		// way speed
+		if ( cnv_max_speed > link.line_way_speed && nexttile.len() > 3 ) {
+
+			local way_speed = 500
+			local upgrade_tiles = 0
+			for ( local i = 0; i < nexttile.len(); i++ ) {
+				local tile_way = tile_x(nexttile[i].x, nexttile[i].y, nexttile[i].z).find_object(mo_way)
+				if ( (tile_way.get_owner().nr == our_player_nr || tile_way.get_owner().nr == 1) ) {
+					upgrade_tiles++
+					if ( tile_way.get_desc().get_topspeed() < way_speed ) {
+						way_speed = tile_way.get_desc().get_topspeed()
+					}
+				}
+
+			}
+			link.line_way_speed = way_speed
+			//gui.add_message_at(our_player, way_speed + " way speed line " + line.get_name(), world.get_time())
+			gui.add_message_at(our_player, upgrade_tiles + " possible tiles for upgrading ", world.get_time())
+			//gui.add_message_at(our_player, " cnv max speed " + cnv_max_speed, world.get_time())
+
+			// upgrade way
+			local way_obj = find_object("way", cnv.get_waytype(), cnv_max_speed)
+			//gui.add_message_at(our_player, " way max speed new " + way_obj.get_topspeed(), world.get_time())
+
+			if ( cnv_max_speed >= way_speed && upgrade_tiles > 2 ) {
+				local costs = (upgrade_tiles*(way_obj.get_cost()/100))
+				if ( our_player.get_current_cash() > costs ) {
+					for ( local i = 1; i < nexttile.len(); i++ ) {
+						local tile_way_1 = tile_x(nexttile[i-1].x, nexttile[i-1].y, nexttile[i-1].z)
+						local tile_way_2 = tile_x(nexttile[i].x, nexttile[i].y, nexttile[i].z)
+						if ( (tile_way_2.find_object(mo_way).get_owner().nr == our_player_nr || tile_way_2.find_object(mo_way).get_owner().nr == 1) && tile_way_2.find_object(mo_way).get_desc().get_topspeed() < way_obj.get_topspeed() ) {
+							command_x.build_way(our_player, tile_way_1, tile_way_2, way_obj, true)
+						}
+					}
+				}
+
+				local start_l = nexttile[nexttile.len()-1]
+				local end_l = nexttile[0]
+
+				local msgtext = format(translate("%s extends the route from %s (%s) to %s (%s)"), our_player.get_name(), start_l.get_halt().get_name(), coord_to_string(start_l), end_l.get_halt().get_name(), coord_to_string(end_l))
+				gui.add_message_at(our_player, msgtext, start_l)
+				//gui.add_message_at(our_player, " upgrade way ", world.get_time())
+
+				link.line_way_speed = way_obj.get_topspeed()
+			} else {
+				// set new way speed by not upgrade way
+				link.line_way_speed = way_obj.get_topspeed()
+			}
+
+		}
+
+		if (cnv.is_valid() && cnv.is_withdrawn()) {
 			// come back later
 			return
 		}
 
 		// try to upgrade
-		if (cnv.has_obsolete_vehicles()  &&  link.next_check < world.get_time().ticks) {
+/*		if (cnv.has_obsolete_vehicles() &&  link.next_check < world.get_time().ticks) {
 			link.next_check = world.get_time().next_month_ticks
 			if (upgrade_link_line(link, line)) {
 				// update successful
+				gui.add_message_at(our_player, "*** upgrade_link_line -> line " + line.get_name(), world.get_time())
 				return
 			}
 		}
-
+*/
 		local lf = link.freight
 		// capacity of convoy
 		local capacity = 0
@@ -338,6 +499,7 @@ class industry_manager_t extends manager_t
 		//local remove_cnv = 0
 		{
 			local list = line.get_convoy_list()
+			local message_show = 0
 			foreach(c in list)
 			{
 				// convoy empty?
@@ -382,19 +544,29 @@ class industry_manager_t extends manager_t
 					//gui.add_message_at(our_player, "####### destroy stucked road vehicles " + cnv_count, world.get_time())
 					c.destroy(our_player)
 					cnv_count--
+					message_show++
 					//remove_cnv++
 				}
-
 			}
+
+			if ( message_show > 0 ) {
+				local msgtext = format(translate("%s removes convoys from line: %s"), our_player.get_name(), line.get_name())
+				gui.add_message_at(our_player, msgtext, world.get_time())
+				return true
+			}
+
 		}
 		dbgprint("Line:  loading = " + cc_load + ", stopped = " + cc_stop + ", new = " + cc_new + ", empty = " + cc_empty)
 		dbgprint("")
 
-		if (freight_available  &&  cc_new == 0  &&  cc_stop < 2) {
+		if (freight_available  &&  cc_new == 0  &&  cc_stop < 2 && cnv.is_valid() ) {
 
 			// stations distance
 			local l = abs(start_l.x - end_l.x) + abs(start_l.y - end_l.y)
 			local c = 0
+			local t = l % 80
+			//gui.add_message_at(our_player, "#### way len " + l + " % 80 = " + (l % 80), world.get_time())
+			//gui.add_message_at(our_player, "#### way len " + l + " / 80 = " + (l / 80), world.get_time())
 			if ( l > 50 && l <= 90 ) {
 				c = 1
 			}	else if ( l > 90 && l <= 160 ) {
@@ -403,10 +575,10 @@ class industry_manager_t extends manager_t
 				c = 3
 			} else if ( l > 220 && l <= 350 ) {
 				c = 4
-			} else if ( l > 350 && l <= 480 ) {
+			} else if ( l > 350 && l < 480 ) {
 				c = 5
-			} else if ( l > 480 ) {
-				c = 6
+			} else if ( l >= 480 ) {
+				c = (l / 80)
 			}
 
 
@@ -435,16 +607,20 @@ class industry_manager_t extends manager_t
 				 *
 				 */
 				local build_double_ways = false
-				if ( s_fields != true ) {
+				if ( s_fields != true && c > 0 && s_fields.len() > 0 ) {
 					local obj_sign = find_signal("is_signal", cnv.get_waytype())
 					local way_obj = s_fields[0].find_object(mo_way).get_desc()
 
 					local build_cost = s_fields.len()*(obj_sign.get_cost()*2)
 					build_cost += s_fields.len()*(way_obj.get_cost()*8)
 
-					// terraform factor
-					build_cost = build_cost*10
-					if ( build_cost < our_player.get_current_cash() ) {
+					// terraform 4 fields
+					build_cost = build_cost+(command_x.slope_get_price(82)*4)
+
+					// count double ways
+					build_cost = build_cost*c
+
+					if ( build_cost/100 < (our_player.get_current_cash()+(our_player.get_current_maintenance()/100*5)) ) {
 						build_double_ways = true
 					}
 
@@ -487,14 +663,10 @@ class industry_manager_t extends manager_t
 					return
 				}
 			} else if (cnv.get_waytype() == wt_rail && (cnv_count > c || c == 0) ) {
-				return
+				return null
 			}
 
-			if ( link.double_ways_count > 0 ) {
-				cnv_count = link.double_ways_count + 1
-				//gui.add_message_at(our_player, "###---- set convoy count from link.double_ways_count " + cnv_count, world.get_time())
-			}
-
+			cnv_count = link.double_ways_count + 1
 
 			if (gain_per_m > 0) {
 				// directly append
@@ -524,13 +696,18 @@ class industry_manager_t extends manager_t
 				cnv_valuator.max_cnvs = 200
 				// no signals and double tracks - limit 1 convoy for rail
 				if (wt == wt_rail) {
-					cnv_valuator.max_cnvs = 1
+					cnv_valuator.max_cnvs = cnv_count
 				}
 
 				// through schedule to estimate distance
 				local dist = 0
 				local entries = cnv.get_schedule().entries
 				dist = abs(entries[0].x - entries[1].x) + abs(entries[0].y - entries[1].y)
+
+				if (wt == wt_road && dist > 40) {
+					cnv_valuator.max_cnvs = dist - 20
+				}
+
 				// add 10% from distance
 				dist += dist / 100 * 10
 
@@ -589,7 +766,7 @@ class industry_manager_t extends manager_t
 
 				local msgtext = format(translate("%s build additional convoy to line: %s"), our_player.get_name(), line.get_name())
 				gui.add_message_at(our_player, msgtext, world.get_time())
-
+				return true
 			}
 		}
 
@@ -597,15 +774,18 @@ class industry_manager_t extends manager_t
 			// freight, lots of empty and of stopped vehicles
 			// -> something is blocked, maybe we block our own supply?
 			// delete one convoy
-			cnv_empty_stopped.destroy(our_player)
-			dbgprint("==> destroy empty convoy")
-			if ( print_message_box == 1 ) {
-				gui.add_message_at(our_player, "####### cnv_count " + cnv_count, world.get_time())
-				gui.add_message_at(our_player, "Line: " + line.get_name() + " ==> destroy empty convoy", world.get_time())
-			}
+			if ( cnv_empty_stopped.is_valid() ) {
+				cnv_empty_stopped.destroy(our_player)
+				dbgprint("==> destroy empty convoy")
+				if ( print_message_box == 1 ) {
+					gui.add_message_at(our_player, "####### cnv_count " + cnv_count, world.get_time())
+					gui.add_message_at(our_player, "Line: " + line.get_name() + " ==> destroy empty convoy", world.get_time())
+				}
 
-			local msgtext = format(translate("%s removes convoys from line: %s"), our_player.get_name(), line.get_name())
-			gui.add_message_at(our_player, msgtext, world.get_time())
+				local msgtext = format(translate("%s removes convoys from line: %s"), our_player.get_name(), line.get_name())
+				gui.add_message_at(our_player, msgtext, world.get_time())
+
+			}
 		}
 		dbgprint("")
 
@@ -617,6 +797,7 @@ class industry_manager_t extends manager_t
 	 */
 	function upgrade_link_line(link, line)
 	{
+		gui.add_message_at(our_player, "### upgrade_link_line " + line.get_name(), world.get_time())
 		// find convoy
 		local cnv = null
 		{
@@ -655,7 +836,7 @@ class industry_manager_t extends manager_t
 			}
 		}
 
-		local wt = wt
+		//local wt = wt
 		// TODO do something smarter
 		prototyper.min_speed  = 1
 		prototyper.max_vehicles = get_max_convoi_length(wt)
@@ -669,6 +850,10 @@ class industry_manager_t extends manager_t
 		cnv_valuator.freight  = link.freight.get_name()
 		cnv_valuator.volume   = transported
 		cnv_valuator.max_cnvs = 200
+		// no signals and double tracks - limit 1 convoy for rail
+		if (wt == wt_rail) {
+			cnv_valuator.max_cnvs = link.double_ways_count + 1
+		}
 		cnv_valuator.distance = dist
 
 		local bound_valuator = valuator_simple_t.valuate_monthly_transport.bindenv(cnv_valuator)
@@ -700,7 +885,11 @@ class industry_manager_t extends manager_t
 		c.p_depot    = depot_x(depot.x, depot.y, depot.z)
 		c.p_line     = line
 		c.p_convoy   = planned_convoy
-		c.p_count    = min(planned_convoy.nr_convoys, 3)
+		if ( wt == wt_rail || wt == wt_water ) {
+			c.p_count    = min(planned_convoy.nr_convoys, 1)
+		} else {
+			c.p_count    = min(planned_convoy.nr_convoys, 3)
+		}
 		c.p_withdraw = true
 		append_child(c)
 		return true
@@ -723,6 +912,39 @@ class industry_manager_t extends manager_t
 	{
 		link_iterator = null // wont save
 		return ::saveinstance("industry_manager_t", this)
+	}
+
+}
+
+function convoy_max_speed(cnv) {
+	// convoy_x::calc_max_speed	(	integer power, integer weight, integer speed_limit )
+
+	local pwr = 0
+	local weight = 0
+	local m_speed = 500
+
+	local veh_list = cnv.get_vehicles()
+
+	for (local i = 0; i < veh_list.len(); i++) {
+		pwr += veh_list[i].get_power()
+		weight += veh_list[i].get_weight() + (veh_list[i].get_capacity()*veh_list[i].get_freight().get_weight_per_unit())
+		if ( m_speed > veh_list[i].get_topspeed() ) {
+			m_speed = veh_list[i].get_topspeed()
+		}
+	}
+
+	local a = cnv.calc_max_speed(pwr, weight, m_speed)
+	return a
+}
+
+function convoy_is_retired(cnv) {
+
+	local veh_list = cnv.get_vehicles()
+
+	for (local i = 0; i < veh_list.len(); i++) {
+		if ( veh_list[i].is_retired(world.get_time()) ) {
+			return true
+		}
 	}
 
 }
